Works nice but at some point the car losses its mind

+++ src/main.cpp
@@ -35,7 +35,7 @@ string hasData(string s) {
 // Evaluate a polynomial.
 double polyeval(Eigen::VectorXd coeffs, double x) {
 	double result = 0.0;
-	for (int i = 0; i < coeffs.size(); i++) {
+	for (int i = 0; i < (int)	coeffs.size(); i++) {
 		result += coeffs[i] * pow(x, i);
 	}
 	return result;
@@ -44,7 +44,7 @@ double polyeval(Eigen::VectorXd coeffs, double x) {
 // Evaluate the derivative of a polynomial.
 double polyeval_derivative(Eigen::VectorXd coeffs, double x) {
 	double result = 0.0;
-	for (int i = 1; i < coeffs.size(); i++) {
+	for (int i = 1; i < (int)coeffs.size(); i++) {
 		result += i * coeffs[i] * pow(x, i-1);
 	}
 	return result;
@@ -53,24 +53,25 @@ double polyeval_derivative(Eigen::VectorXd coeffs, double x) {
 // Adapted from
 // https://github.com/JuliaMath/Polynomials.jl/blob/master/src/Polynomials.jl#L676-L716
 Eigen::VectorXd polyfit(Eigen::VectorXd xvals, Eigen::VectorXd yvals,
-                        int order) {
-  assert(xvals.size() == yvals.size());
-  assert(order >= 1 && order <= xvals.size() - 1);
-  Eigen::MatrixXd A(xvals.size(), order + 1);
+                        int order)
+{
+	assert(xvals.size() == yvals.size());
+	assert(order >= 1 && order <= xvals.size() - 1);
+	Eigen::MatrixXd A(xvals.size(), order + 1);
 
-  for (int i = 0; i < xvals.size(); i++) {
-    A(i, 0) = 1.0;
-  }
+	for (int i = 0; i < xvals.size(); i++) {
+		A(i, 0) = 1.0;
+	}
 
-  for (int j = 0; j < xvals.size(); j++) {
-    for (int i = 0; i < order; i++) {
-      A(j, i + 1) = A(j, i) * xvals(j);
-    }
-  }
+	for (int j = 0; j < xvals.size(); j++) {
+		for (int i = 0; i < order; i++) {
+			A(j, i + 1) = A(j, i) * xvals(j);
+		}
+	}
 
-  auto Q = A.householderQr();
-  auto result = Q.solve(yvals);
-  return result;
+	auto Q = A.householderQr();
+	auto result = Q.solve(yvals);
+	return result;
 }
